feat(burper): fix burper not being able to parse files feat(burp): change  the way environment variables are handled
Environment variables are now handled either through client-side (CLI) or the server-side, which are indicated  in the burp.toml

This change is introduced to fix an issue with transporting secrets from environment variables. By default, server-side is disabled because  the secrets tends to lies in the developer's environment (deployer), but in the case where the server has the secrets, you can enable server-side instead.